requirements.txt
django==2.0.3
djangorestframework==3.9.4
Pillow==7.2.0
psycopg2==2.8.5